gate/requirements.txt
aiohttp>=3.6.2
aioredis>=1.3.1
asyncio-nats-client>=0.10.0
